django-cognito/requirements.txt
django==3.1.4
django-rest-framework==0.1.0
django-filter==2.3.0
django-environ==0.4.5
django-cors-headers==3.4.0
pyjwt==1.7.1
drf-jwt==1.19.1
boto3==1.16.47
prettytable
orbitkit==0.4.6
